make it generic to ALL cached language systems incase it starts being used by other systems in the future.
